+++ resolwe_bio/processes/variant_calling/genotype_gvcfs.py
@@ -1,4 +1,7 @@
 """Run GATK GenotypeGVCFs tool."""
+from pathlib import Path
+
+from joblib import Parallel, delayed, wrap_non_picklable_objects
 from plumbum import TEE
 
 from resolwe.process import (
@@ -8,11 +11,46 @@ from resolwe.process import (
     FileField,
     GroupField,
     IntegerField,
-    ListField,
     Process,
     SchedulingClass,
     StringField,
 )
+from resolwe.process.fields import DirField
+
+
+def create_vcf_path(interval_path):
+    """Create a vcf path based on the interval name."""
+    return f"cohort_variants/{interval_path.stem}.vcf"
+
+
+@delayed
+@wrap_non_picklable_objects
+def run_genotype_gvcfs(interval_path, ref_seq_path, db_path, dbsnp_path):
+    """Run genotyping on a specifed interval."""
+    variants_interval = create_vcf_path(interval_path)
+
+    genotype_gvcfs_inputs = [
+        "-R",
+        ref_seq_path,
+        "-V",
+        f"gendb://{db_path}",
+        "-O",
+        variants_interval,
+        "-L",
+        interval_path,
+        "-D",
+        dbsnp_path,
+        "-G",
+        "StandardAnnotation",
+        "-G",
+        "AS_StandardAnnotation",
+        "--only-output-calls-starting-in-intervals",
+    ]
+
+    return_code, _, _ = Cmd["gatk"]["GenotypeGVCFs"][genotype_gvcfs_inputs] & TEE(
+        retcode=None
+    )
+    return return_code
 
 
 class GatkGenotypeGVCFs(Process):
@@ -22,7 +60,7 @@ class GatkGenotypeGVCFs(Process):
     name = "GATK GenotypeGVCFs"
     category = "GATK"
     process_type = "data:variants:vcf:genotypegvcfs"
-    version = "1.0.1"
+    version = "2.0.0"
     scheduling_class = SchedulingClass.BATCH
     requirements = {
         "expression-engine": "jinja",
@@ -30,8 +68,9 @@ class GatkGenotypeGVCFs(Process):
             "docker": {"image": "public.ecr.aws/s4q6j6e8/resolwebio/dnaseq:6.0.0"}
         },
         "resources": {
-            "cores": 4,
+            "cores": 16,
             "memory": 32768,
+            "storage": 200,
         },
     }
     data_name = "Cohort variants"
@@ -39,16 +78,8 @@ class GatkGenotypeGVCFs(Process):
     class Input:
         """Input fields for GatkGenotypeGVCFs."""
 
-        gvcfs = ListField(
-            DataField("variants:gvcf"),
-            label="Input data (GVCF)",
-        )
+        database = DataField("genomicsdb", label="GATK GenomicsDB")
         ref_seq = DataField("seq:nucleotide", label="Reference sequence")
-
-        intervals = DataField(
-            "bed",
-            label="Intervals file (.bed)",
-        )
 
         dbsnp = DataField("variants:vcf", label="dbSNP file")
 
@@ -88,6 +119,13 @@ class GatkGenotypeGVCFs(Process):
                 "have significantly higher overheads. This flag has no "
                 "effect if only one batch is used.",
             )
+            n_jobs = IntegerField(
+                label="Number of concurent jobs",
+                description="Use a fixed number of jobs for genotyping "
+                "instead of determining it based on the number of available "
+                "cores.",
+                required=False,
+            )
 
         advanced_options = GroupField(
             AdvancedOptions, label="Advanced options", hidden="!advanced"
@@ -97,6 +135,7 @@ class GatkGenotypeGVCFs(Process):
         """Output fields for GatkGenotypeGVCFs."""
 
         vcf = FileField(label="GVCF file")
+        vcf_dir = DirField(label="Folder with split GVCFs")
         tbi = FileField(label="Tabix index")
         species = StringField(label="Species")
         build = StringField(label="Build")
@@ -106,66 +145,70 @@ class GatkGenotypeGVCFs(Process):
         variants = "cohort_variants.vcf"
         variants_gz = variants + ".gz"
         variants_index = variants_gz + ".tbi"
+        intervals_path = Path("intervals_folder")
 
-        sample_map_file = "sample_map.txt"
+        Path("cohort_variants").mkdir(exist_ok=True)
+        intervals_path.mkdir(exist_ok=True)
 
-        species = inputs.gvcfs[0].output.species
-        if not all(gvcf.output.species == species for gvcf in inputs.gvcfs):
-            self.error("Not all of the input samples are of the same species.")
+        if inputs.advanced_options.n_jobs:
+            n_jobs = max(inputs.advanced_options.n_jobs, 1)
+        else:
+            n_jobs = max(self.requirements.resources.cores, 1)
 
-        build = inputs.gvcfs[0].output.build
-        if not all(gvcf.output.build == build for gvcf in inputs.gvcfs):
-            self.error("Not all of the input samples have the same genome build.")
-
-        with open(sample_map_file, "w") as sample_map:
-            for gvcf in inputs.gvcfs:
-                sample_map.write(f"{gvcf.entity_name}\t{gvcf.output.vcf.path}\n")
-
-        db_import_args = [
-            "--genomicsdb-workspace-path",
-            "database",
-            "-L",
-            inputs.intervals.output.bed.path,
-            "--sample-name-map",
-            sample_map_file,
-            "--batch-size",
-            inputs.advanced_options.batch_size,
-            "--reader-threads",
-            min(self.requirements.resources.cores, 5),
-        ]
-
-        if inputs.advanced_options.consolidate:
-            db_import_args.append("--seqBias")
-
-        return_code, _, _ = Cmd["gatk"]["GenomicsDBImport"][db_import_args] & TEE(
-            retcode=None
-        )
-        if return_code:
-            self.error("GATK GenomicsDBImport tool failed.")
-
-        genotype_gvcfs_inputs = [
+        split_intervals_inputs = [
             "-R",
             inputs.ref_seq.output.fasta.path,
-            "-V",
-            "gendb://database",
-            "-O",
-            variants,
             "-L",
-            inputs.intervals.output.bed.path,
-            "-D",
-            inputs.dbsnp.output.vcf.path,
-            "-G",
-            "StandardAnnotation",
-            "-G",
-            "AS_StandardAnnotation",
-            "--only-output-calls-starting-in-intervals",
+            inputs.database.output.intervals.path,
+            "--scatter-count",
+            n_jobs,
+            "-O",
+            str(intervals_path),
         ]
-
-        return_code, _, _ = Cmd["gatk"]["GenotypeGVCFs"][genotype_gvcfs_inputs] & TEE(
+        return_code, _, _ = Cmd["gatk"]["SplitIntervals"][split_intervals_inputs] & TEE(
             retcode=None
         )
-        if return_code:
+
+        if return_code is None:
+            self.error("Could not create equally sized intervals from the bedfile.")
+
+        intervals = [path for path in intervals_path.glob("*.interval_list")]
+
+        return_codes = Parallel(n_jobs=n_jobs)(
+            run_genotype_gvcfs(
+                interval_path,
+                ref_seq_path=inputs.ref_seq.output.fasta.path,
+                db_path=inputs.database.output.database.path,
+                dbsnp_path=inputs.dbsnp.output.vcf.path,
+            )
+            for interval_path in intervals
+        )
+
+        outputs.vcf_dir = "cohort_variants"
+        if any(return_codes):
             self.error("GATK GenotypeGVCFs tool failed.")
+
+        merge_list_file = "merge_files.list"
+        with open(merge_list_file, "w") as f:
+            for interval_path in sorted(intervals):
+                f.write(f"{create_vcf_path(interval_path)}\n")
+
+        self.progress(0.8)
+
+        merge_inputs = [
+            "-I",
+            merge_list_file,
+            "-O",
+            variants,
+            "-R",
+            inputs.ref_seq.output.fasta.path,
+            "--CREATE_INDEX",
+            "false",
+        ]
+
+        return_code, _, _ = Cmd["gatk"]["GatherVcfs"][merge_inputs] & TEE(retcode=None)
+        if return_code:
+            self.error("GATK GatherVcfs tool failed.")
 
         # Compress and index the output variants file
         (Cmd["bgzip"]["-c", variants] > variants_gz)()
@@ -173,5 +216,5 @@ class GatkGenotypeGVCFs(Process):
 
         outputs.vcf = variants_gz
         outputs.tbi = variants_index
-        outputs.species = species
-        outputs.build = build
+        outputs.species = inputs.database.output.species
+        outputs.build = inputs.database.output.build