+++ resolwe_bio/tests/processes/test_wgs.py
@@ -244,6 +244,114 @@ class WgsProcessorTestCase(BioProcessTestCase):
         self.assertFields(variants, "build", "custom_build")
         self.assertFields(variants, "species", "Homo sapiens")
 
+    @tag_process("gatk-genomicsdb-import")
+    def test_gatk_genomicsdb(self):
+        base = Path("wgs")
+        inputs = base / "input"
+        with self.preparation_stage():
+            intervals = self.run_process(
+                "upload-bed",
+                {
+                    "src": inputs / "hg38.intervals.bed",
+                    "species": "Homo sapiens",
+                    "build": "custom_build",
+                },
+            )
+
+            # Mock upload gvcf process
+            process = Process.objects.create(
+                name="Upload GVCF mock process",
+                requirements={
+                    "expression-engine": "jinja",
+                    "resources": {
+                        "network": True,
+                    },
+                    "executor": {
+                        "docker": {
+                            "image": "public.ecr.aws/s4q6j6e8/resolwebio/base:ubuntu-20.04-03042021",
+                        },
+                    },
+                },
+                contributor=self.contributor,
+                type="data:variants:gvcf:",
+                entity_type="sample",
+                entity_descriptor_schema="sample",
+                data_name="{{ gvcf.file }}",
+                input_schema=[
+                    {
+                        "name": "gvcf",
+                        "type": "basic:file:",
+                    },
+                    {
+                        "name": "tabix",
+                        "type": "basic:file:",
+                    },
+                ],
+                output_schema=[
+                    {
+                        "name": "vcf",
+                        "type": "basic:file:",
+                    },
+                    {
+                        "name": "tbi",
+                        "type": "basic:file:",
+                    },
+                    {
+                        "name": "species",
+                        "type": "basic:string:",
+                    },
+                    {
+                        "name": "build",
+                        "type": "basic:string:",
+                    },
+                ],
+                run={
+                    "language": "bash",
+                    "program": r"""
+re-import {{ gvcf.file_temp|default(gvcf.file) }} {{ gvcf.file }} "g.vcf" "g.vcf" 0.1 compress
+re-save-file vcf "${NAME}.g.vcf.gz"
+re-import {{ tabix.file_temp|default(tabix.file) }} {{ tabix.file }} "g.vcf.gz.tbi" "g.vcf.gz.tbi" 0.1 extract
+re-save-file tbi "${NAME}.g.vcf.gz.tbi"
+re-save species "Homo sapiens"
+re-save build "custom_build"
+""",
+                },
+            )
+
+            gvcf_input = {
+                "gvcf": inputs / "variants.g.vcf.gz",
+                "tabix": inputs / "variants.g.vcf.gz.tbi",
+            }
+            gvcf_1 = self.run_process(process.slug, gvcf_input)
+
+            gvcf_input = {
+                "gvcf": inputs / "variants2.g.vcf.gz",
+                "tabix": inputs / "variants2.g.vcf.gz.tbi",
+            }
+            gvcf_2 = self.run_process(process.slug, gvcf_input)
+
+        database = self.run_process(
+            "gatk-genomicsdb-import",
+            {
+                "gvcfs": [gvcf_1.id],
+                "intervals": intervals.id,
+            },
+        )
+
+        self.assertFields(database, "build", "custom_build")
+        self.assertFields(database, "species", "Homo sapiens")
+
+        database2 = self.run_process(
+            "gatk-genomicsdb-import",
+            {
+                "gvcfs": [gvcf_2.id],
+                "use_existing": True,
+                "existing_db": database.id,
+            },
+        )
+        self.assertFields(database2, "build", "custom_build")
+        self.assertFields(database2, "species", "Homo sapiens")
+
     @tag_process("gatk-genotype-gvcfs")
     def test_gatk_genotypegvcfs(self):
         base = Path("wgs")
@@ -349,12 +457,19 @@ re-save build "custom_build"
             }
             gvcf_2 = self.run_process(process.slug, gvcf_input)
 
+            database = self.run_process(
+                "gatk-genomicsdb-import",
+                {
+                    "gvcfs": [gvcf_1.id, gvcf_2.id],
+                    "intervals": intervals.id,
+                },
+            )
+
         joint_variants = self.run_process(
             "gatk-genotype-gvcfs",
             {
-                "gvcfs": [gvcf_1.id, gvcf_2.id],
+                "database": database.id,
                 "ref_seq": ref_seq.id,
-                "intervals": intervals.id,
                 "dbsnp": dbsnp.id,
             },
         )