+++ docs/CHANGELOG.rst
@@ -25,6 +25,9 @@ Changed
   FASTQ files
 - **BACKWARD INCOMPATIBLE:** Rewrite ``feature_counts`` process to
   Python
+- **BACKWARD INCOMPATIBLE:** Create a separate process
+  ``gatk-genomicsdb-import`` for importing GVCFs into the database and
+  use it as an input in the parallelised ``gatk-genotype-gvcfs`` process
 - Support ChIP-Seq and ATAC-Seq data sets in ``geo-import`` process
 
 Fixed